+++ core/src/main/java/io/cucumber/core/runner/TestCaseState.java
@@ -68,8 +68,14 @@ class TestCaseState implements io.cucumber.core.backend.TestCaseState {
         embed(data, mediaType, null);
     }
 
+    @Deprecated
     @Override
     public void embed(byte[] data, String mediaType, String name) {
+        attach(data, mediaType, name);
+    }
+
+    @Override
+    public void attach(byte[] data, String mediaType, String name) {
         bus.send(new EmbedEvent(bus.getInstant(), testCase, data, mediaType, name));
         bus.send(Messages.Envelope.newBuilder()
             .setAttachment(
@@ -86,8 +92,14 @@ class TestCaseState implements io.cucumber.core.backend.TestCaseState {
         );
     }
 
+    @Deprecated
     @Override
     public void write(String text) {
+        log(text);
+    }
+
+    @Override
+    public void log(String text) {
         bus.send(new WriteEvent(bus.getInstant(), testCase, text));
         bus.send(Messages.Envelope.newBuilder()
             .setAttachment(